refactor(runtime): merge runtime into core
