+++ crates/core/src/lib.rs
@@ -6,6 +6,7 @@
 //! - [`Dispatcher`]: Generic async trait for tool dispatch.
 //! - [`ToolRegistry`]: Canonical dispatcher — holds `(Tool, Handler)` pairs.
 //! - [`Hook`]: Lifecycle backend for agent building, events, and tool registration.
+//! - [`Runtime`]: Agent registry, tool registry, and hook orchestration.
 //! - [`model`]: Unified LLM interface types and traits.
 //! - Agent event types: [`AgentEvent`], [`AgentStep`], [`AgentResponse`], [`AgentStopReason`].
 
@@ -15,7 +16,7 @@ pub use agent::{
     parse_agent_md,
     tool::{Dispatcher, Handler, ToolRegistry},
 };
-pub use runtime::hook::Hook;
+pub use runtime::{Runtime, hook::Hook};
 
 mod agent;
 pub mod model;
