+++ crates/daemon/Cargo.toml
@@ -23,7 +23,6 @@ memory.workspace = true
 model.workspace = true
 protocol.workspace = true
 router.workspace = true
-runtime.workspace = true
 skill.workspace = true
 socket.workspace = true
 wcore.workspace = true
