+++ crates/daemon/src/daemon/builder.rs
@@ -7,8 +7,8 @@ use crate::{
 };
 use anyhow::Result;
 use model::ProviderManager;
-use runtime::Runtime;
 use std::path::{Path, PathBuf};
+use wcore::Runtime;
 
 /// Step-by-step builder for the daemon's [`Runtime`].
 ///
