--- crates/runtime/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "walrus-runtime"
-description = "Walrus runtime"
-documentation = "https://docs.rs/walrus-runtime"
-readme = "README.md"
-version.workspace = true
-edition.workspace = true
-authors.workspace = true
-license.workspace = true
-repository.workspace = true
-keywords.workspace = true
-
-[dependencies]
-memory.workspace = true
-wcore.workspace = true
-
-# crates.io
-anyhow.workspace = true
-async-stream.workspace = true
-compact_str.workspace = true
-futures-core.workspace = true
-futures-util.workspace = true
-tokio.workspace = true
-tracing.workspace = true
